docs: Use a single value for algolia version meta
As-is we have an array like `["v0.13.4", "latest"]` but it seems that
we aren't referencing it as such, its trying to compare a string
version.

It looks like on the `latest` urls we would set the facetFilter to
be the explicit version, so we'll just stick with that for the meta
tags too.

Signed-off-by: Patrick East <east.patrick@gmail.com>