cmd: Add `-b`/`--bundle` option to opa run
This adds two new flags to match the ones on test, build, and eval
where it will load files/directories as bundles.

As some point in the future we might want to restrict the ability
to load bundles (or >1 bundle) with the normal non-bundle load path
as it doesn't actually work as expected.. For now this doesn't make
any changes to that behavior.

Signed-off-by: Patrick East <east.patrick@gmail.com>