docs: Add links to homepage
Previously, there were no links from the homepage to any of the
more detailed docs about use cases and even mention of use cases
was below the fold.

This change adds links to use cases at the very top of the page,
along with pointers to core docs and the OPA summit.

Signed-off-by: Tim Hinrichs <tim@styra.com>