docs: Use full copy for `latest` instead of shadow
Instead of using url rewriting tricks we will just copy the latest
docs as part of the build process for a "latest" version.

In the version selector code we need to handle this to hide the
version, but overall it simplifies the template code.. so thats nice.

Best part, and main reason for doing this, is that the pretty url
feature will now work on the `/docs/latest/*` urls. It will fix some
weirdness with the relative url links we use in the markdown content
and remove a duplicate search result.

Signed-off-by: Patrick East <east.patrick@gmail.com>