docs: Add redirect for /docs/latest/get-started
In the re-org we got rid of the getting started section and instead
have a more comprehensive/interactive introduction page. This change
adds a redirect to send /docs/latest/get-started URLs to /docs/latest
instead.

We also remove the link to it in the custom 404 page. Its the same
as the documentation URL now.

Signed-off-by: Patrick East <east.patrick@gmail.com>