Fix *-debug docker images to be ":debug" tag based
We use the gcr.io/distroless/base:debug tag for the *-debug images of
OPA. At the point where we switched to the multi-stage build we dropped
defining the `VARIANT` argument so it wasn't being picked up correctly.

This lead to us just using gcr.io/distroless/base for the *-debug
image.. which isn't what we wanted.

Signed-off-by: Patrick East <east.patrick@gmail.com>