wasm: Refactor to pass eval() parsed input
Previously the wasm backend would emit code to parse the raw input
document inside of hte eval() function. This commit updates the eval()
function to expect parsed input. This change is being made because
we're going to end up with parsing and address space setup happening
in the host environment for cached data/context.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>