docs: Preprocess live-blocks with current OPA HEAD
This changes to no longer download the `edge` release binary (we still
publish one for anyone who wants to use the current master). This way
for PR's and for development we are going to process the `edge`
documentation with the version of OPA that is the current HEAD. For
the live site this will always be the current master branch.

Signed-off-by: Patrick East <east.patrick@gmail.com>