Add API to disable rule indexing for evaluation
For debugging we would like to have the ability to retrieve all rules
from the Index rather than the filtered ruleset. This API will simply
walk the tree and collect all of the rules.

This allows a client making evaluations to disable the rule indexing,
which particularly useful for debugging purposes.

Essentially this boils down to just calling `AllRules()` versus using
the `Lookup()` API on the index.

There is plumbing included to pass the flag in from the rego API's
down through to the topdown eval.

Signed-off-by: Patrick East <east.patrick@gmail.com>