Add debug tracing mode
This adds a new option for eval which will disable indexing so
that variable bindings for rules that would otherwise not evaluate
can be found.

This also adds to (and corrects) the JSON marshalled trace events. We
now have extra metadata about the local variables, a working JSON
marshaller for the `Locals` (which includes type info), and the event
nodes location information.

Fixes: #1697
Signed-off-by: Patrick East <east.patrick@gmail.com>