wasm: Fix bug in reference planning
The target local was not being set correctly before recursing on the
reference. If the reference operand was a non-ground composite (e.g.,
p[[1,x]]) then the target would get clobbered by the last ground term
in the composite to be planned (1 in that example). This was not
noticed during testing because we rarely dereference set elements
while iterating over them (e.g., p[[1,x]][1]).

Signed-off-by: Torin Sandall <torinsandall@gmail.com>