ast: Fix object corruption during local rewrite
When rewriting declared vars in objects the compiler was not rewriting
the object keys recursively during construction of the new
object. This meant the new object's hashtable would get corrupted when
the outer visitor ultimately performed the rewriting on ref keys.

Since the compiler already keeps track of whether the outer visitor
needs to recurse, the fix simply updates the compiler to rewrite
object key/value terms recursively when the new object is being
created and stops the outer visitor from recursing on the new object.

Sets are given the same treatment because they have the same
underlying hashtable (though it's unlikely we would see the issue for
sets since the hashtable is only used for Contains() and we rarely, if
ever, call Contains() w/ var terms).

Fixes #1852

Signed-off-by: Torin Sandall <torinsandall@gmail.com>