wasm: Add WithStmt to IR and implement on backend
This change updates the compiler to support a WithStmt instruction in
the IR. The WithStmt has the same upsert behaviour as in Rego however
the statement returns the modified value as opposed to modifying it in
place. The WithStmt will allow the planner to easily insert values
into the input and data documents. In theory, the WithStmt could also
be used on locals.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>