topdown: Invoke iterator when evaluating negation
In `evalNot` when it succeeds we were calling `e.next(..)` to continue
evaluation, but if the previous expression was from `evalWith` the
original state was not being restored. This caused the patched data
to persist for subsequent expression evaluation(s) until `iter()` was
used.

Similar treatment was required in `evalNotPartial` for the partial
evaluation case.

Fixes: #2142
Signed-off-by: Patrick East <east.patrick@gmail.com>