docs: Force netlify redirects
Coming in the next few weeks the old redirects defined in `_redirects`
will not behave the same. To keep them working the same as before we
need to tell Netlify to force them (rather than looking for shadowed
pages). By specifying the return code and adding a `!` we achieve the
original behavior.

Signed-off-by: Patrick East <east.patrick@gmail.com>